Attempt to put GeometryExtension in the Mesh.
